#noid #done #comment [protects 에러 처리]

See merge request s08-webmobile1-sub2/S08P12E104!130

+++ backend/src/main/java/com/ssafy/patpat/shelter/dto/ShelterDto.java
@@ -31,5 +31,5 @@ public class ShelterDto {
     private String gugunCode;
     private String gugunName;
     private String adminName;
-    private int adminId;
+    private Long adminId;
 }
